finalized adding fmt library
