+++ src/tnt/util/All.hpp
@@ -5,8 +5,9 @@
 #include "tnt/util/container/ExtendHash.hpp"
 #include "tnt/util/iterables/KeysValues.hpp"
 #include "tnt/util/Types.hpp"
+#include "tnt/util/FmtHelper.hpp"
 #include <itertools.hpp>
-#include <fmt/format.h>
+
 
 using namespace iter;
